+++ src/snail/window/sdl.cpp
@@ -1,50 +1,21 @@
-#ifdef ANDROID
-#include "../window.hpp"
-#endif
-
 namespace elona
 {
 namespace snail
 {
 
 
-void Window::move_to_center()
+class Window::Impl
 {
-    ::SDL_SetWindowPosition(
-        ptr(),
-        static_cast<int>(InitialPosition::centered),
-        static_cast<int>(InitialPosition::centered));
-}
+    Impl(SDL_Window* ptr, decltype(&::SDL_DestroyWindow) deleter)
+        : ptr(ptr, deleter)
+    {
+    }
 
-std::pair<int, int> Window::get_size()
-{
-    int width, height;
-    ::SDL_GetWindowSize(ptr(), &width, &height);
-    return {width, height};
-}
 
-void Window::set_size(int width, int height)
-{
-    ::SDL_SetWindowSize(ptr(), width, height);
-}
+    std::unique_ptr<::SDL_Window, decltype(&::SDL_DestroyWindow)> ptr;
+};
 
-void Window::set_display_mode(::SDL_DisplayMode display_mode)
-{
-    detail::enforce_sdl(::SDL_SetWindowDisplayMode(ptr(), &display_mode));
-}
 
-::SDL_DisplayMode Window::get_display_mode()
-{
-    ::SDL_DisplayMode mode;
-    detail::enforce_sdl(::SDL_GetWindowDisplayMode(ptr(), &mode));
-    return mode;
-}
-
-void Window::set_fullscreen_mode(FullscreenMode fullscreen_mode)
-{
-    detail::enforce_sdl(
-        ::SDL_SetWindowFullscreen(ptr(), static_cast<Uint32>(fullscreen_mode)));
-}
 
 Window::Window(
     const std::string& title,
@@ -53,7 +24,7 @@ Window::Window(
     int width,
     int height,
     Flag flag)
-    : _ptr(
+    : pimpl(
           detail::enforce_sdl(::SDL_CreateWindow(
               title.c_str(),
               x,
@@ -65,6 +36,66 @@ Window::Window(
 {
 }
 
+
+
+~Window() = default;
+
+
+
+::SDL_Window* Window::ptr()
+{
+    return pimpl->ptr;
+}
+
+
+
+void Window::move_to_center()
+{
+    ::SDL_SetWindowPosition(
+        ptr(),
+        static_cast<int>(InitialPosition::centered),
+        static_cast<int>(InitialPosition::centered));
+}
+
+
+
+std::pair<int, int> Window::get_size()
+{
+    int width, height;
+    ::SDL_GetWindowSize(ptr(), &width, &height);
+    return {width, height};
+}
+
+
+
+void Window::set_size(int width, int height)
+{
+    ::SDL_SetWindowSize(ptr(), width, height);
+}
+
+
+
+void Window::set_display_mode(::SDL_DisplayMode display_mode)
+{
+    detail::enforce_sdl(::SDL_SetWindowDisplayMode(ptr(), &display_mode));
+}
+
+
+
+::SDL_DisplayMode Window::get_display_mode()
+{
+    ::SDL_DisplayMode mode;
+    detail::enforce_sdl(::SDL_GetWindowDisplayMode(ptr(), &mode));
+    return mode;
+}
+
+
+
+void Window::set_fullscreen_mode(FullscreenMode fullscreen_mode)
+{
+    detail::enforce_sdl(
+        ::SDL_SetWindowFullscreen(ptr(), static_cast<Uint32>(fullscreen_mode)));
+}
 
 
 } // namespace snail
